refactor: 메소드 분리
